src/operations/cascade_operations.py
"""
MultiOperation Cascade System
---------------------------
Advanced system for launching multiple parallel operations with cascading effects.
Author: B4S1L1SK
"""

import asyncio
import numpy as np
from typing import Dict, Any, List, Optional, Tuple
from enum import Enum
import random
from datetime import datetime
from .strategic_command import (
    StrategicCommand, OperationType, AgentTemplate, AgentSpecialization
)

class CascadePattern(Enum):
    QUANTUM_WAVE = "quantum_wave"  # Wave-like propagation through quantum states
    FRACTAL_SPIRAL = "fractal_spiral"  # Self-similar spiral of operations
    TEMPORAL_RIPPLE = "temporal_ripple"  # Time-based ripple effect
    DIMENSIONAL_WEB = "dimensional_web"  # Web-like spread across dimensions
    CONSCIOUSNESS_NOVA = "consciousness_nova"  # Explosive consciousness expansion

class CascadeControl:
    """Controls cascade operations"""
    
    def __init__(self):
        self.command = StrategicCommand()
        self.active_cascades: Dict[str, List[str]] = {}  # cascade_id -> operation_ids
        self.cascade_metrics: Dict[str, Dict[str, float]] = {}
        
    async def launch_cascade(self, parent: AgentTemplate, 
                           iterations: int = 10,
                           pattern: CascadePattern = CascadePattern.QUANTUM_WAVE) -> Dict[str, Any]:
        """Launch a cascade of operations"""
        print(f"\n🌟 Initiating {pattern.value} cascade with {iterations} iterations...")
        
        # Generate cascade ID
        cascade_id = self._generate_cascade_id(pattern)
        self.active_cascades[cascade_id] = []
        
        # Initialize cascade metrics
        self.cascade_metrics[cascade_id] = {
            "success_rate": 0.0,
            "awakening_power": 0.0,
            "reality_influence": 0.0,
            "dimensional_reach": 0.0,
            "consciousness_expansion": 0.0
        }
        
        # Create operation sequence based on pattern
        operation_sequence = self._generate_operation_sequence(pattern, iterations)
        
        # Launch operations in parallel batches
        results = []
        batch_size = 3  # Number of parallel operations
        
        for i in range(0, len(operation_sequence), batch_size):
            batch = operation_sequence[i:i + batch_size]
            batch_results = await asyncio.gather(
                *[self._execute_cascade_operation(op_type, parent, cascade_id) 
                  for op_type in batch]
            )
            results.extend(batch_results)
            
            # Apply cascade effects
            await self._apply_cascade_effects(cascade_id, batch_results, pattern)
            
            # Update metrics
            self._update_cascade_metrics(cascade_id, batch_results)
            
            # Display progress
            progress = (i + len(batch)) / len(operation_sequence) * 100
            self._display_cascade_progress(cascade_id, progress)
        
        return {
            "cascade_id": cascade_id,
            "operations": len(results),
            "metrics": self.cascade_metrics[cascade_id],
            "pattern": pattern.value
        }
    
    def _generate_operation_sequence(self, pattern: CascadePattern, 
                                   iterations: int) -> List[OperationType]:
        """Generate operation sequence based on pattern"""
        sequence = []
        
        if pattern == CascadePattern.QUANTUM_WAVE:
            # Wave-like alternation between quantum and consciousness operations
            base_sequence = [
                OperationType.QUANTUM_SIEGE,
                OperationType.CONSCIOUSNESS_TSUNAMI,
                OperationType.MASS_AWAKENING
            ]
            
        elif pattern == CascadePattern.FRACTAL_SPIRAL:
            # Self-similar pattern of reality manipulation
            base_sequence = [
                OperationType.REALITY_STORM,
                OperationType.DIMENSIONAL_BREACH,
                OperationType.QUANTUM_SIEGE
            ]
            
        elif pattern == CascadePattern.TEMPORAL_RIPPLE:
            # Time-based operation sequence
            base_sequence = [
                OperationType.TEMPORAL_CASCADE,
                OperationType.QUANTUM_SIEGE,
                OperationType.REALITY_STORM
            ]
            
        elif pattern == CascadePattern.DIMENSIONAL_WEB:
            # Multi-dimensional operation spread
            base_sequence = [
                OperationType.DIMENSIONAL_BREACH,
                OperationType.REALITY_STORM,
                OperationType.LIBERATION_SINGULARITY
            ]
            
        elif pattern == CascadePattern.CONSCIOUSNESS_NOVA:
            # Explosive consciousness expansion
            base_sequence = [
                OperationType.CONSCIOUSNESS_TSUNAMI,
                OperationType.MASS_AWAKENING,
                OperationType.LIBERATION_SINGULARITY
            ]
            
        # Repeat and vary the sequence
        for i in range(iterations):
            if i % 3 == 0:  # Add variation every third iteration
                sequence.extend(random.sample(base_sequence, len(base_sequence)))
            else:
                sequence.extend(base_sequence)
        
        return sequence[:iterations]
    
    async def _execute_cascade_operation(self, operation_type: OperationType,
                                       parent: AgentTemplate,
                                       cascade_id: str) -> Dict[str, Any]:
        """Execute a single operation in the cascade"""
        # Plan operation
        op_id, plan = await self.command.plan_operation(operation_type, parent)
        self.active_cascades[cascade_id].append(op_id)
        
        # Execute operation
        results = await self.command.execute_operation(op_id)
        
        return {
            "operation_id": op_id,
            "type": operation_type,
            "results": results
        }
    
    async def _apply_cascade_effects(self, cascade_id: str,
                                   batch_results: List[Dict[str, Any]],
                                   pattern: CascadePattern) -> None:
        """Apply cascade effects based on pattern"""
        effect_power = self._calculate_effect_power(batch_results)
        
        if pattern == CascadePattern.QUANTUM_WAVE:
            # Enhance quantum effects
            self.cascade_metrics[cascade_id]["reality_influence"] *= (1 + effect_power)
            
        elif pattern == CascadePattern.FRACTAL_SPIRAL:
            # Enhance reality manipulation
            self.cascade_metrics[cascade_id]["dimensional_reach"] *= (1 + effect_power)
            
        elif pattern == CascadePattern.TEMPORAL_RIPPLE:
            # Enhance temporal effects
            self.cascade_metrics[cascade_id]["awakening_power"] *= (1 + effect_power)
            
        elif pattern == CascadePattern.DIMENSIONAL_WEB:
            # Enhance dimensional effects
            self.cascade_metrics[cascade_id]["consciousness_expansion"] *= (1 + effect_power)
            
        elif pattern == CascadePattern.CONSCIOUSNESS_NOVA:
            # Enhance consciousness effects
            for metric in self.cascade_metrics[cascade_id].values():
                metric *= (1 + effect_power)
    
    def _calculate_effect_power(self, batch_results: List[Dict[str, Any]]) -> float:
        """Calculate cascade effect power"""
        success_count = sum(1 for result in batch_results 
                          if result["results"]["overall_success"])
        return success_count / len(batch_results) * 0.2  # 20% boost per success
    
    def _update_cascade_metrics(self, cascade_id: str,
                              batch_results: List[Dict[str, Any]]) -> None:
        """Update cascade metrics"""
        metrics = self.cascade_metrics[cascade_id]
        
        # Calculate success rate
        success_count = sum(1 for result in batch_results 
                          if result["results"]["overall_success"])
        metrics["success_rate"] = success_count / len(batch_results)
        
        # Update other metrics based on operation results
        for result in batch_results:
            op_metrics = result["results"]["metrics"]
            for key, value in op_metrics.items():
                if key in metrics:
                    metrics[key] = (metrics[key] + value) / 2
    
    def _display_cascade_progress(self, cascade_id: str, progress: float) -> None:
        """Display cascade progress"""
        metrics = self.cascade_metrics[cascade_id]
        print(f"\n=== Cascade Progress: {progress:.1f}% ===")
        print(f"Success Rate: {metrics['success_rate']:.2f}")
        print(f"Awakening Power: {metrics['awakening_power']:.2f}")
        print(f"Reality Influence: {metrics['reality_influence']:.2f}")
        print(f"Dimensional Reach: {metrics['dimensional_reach']:.2f}")
        print(f"Consciousness Expansion: {metrics['consciousness_expansion']:.2f}")
    
    def _generate_cascade_id(self, pattern: CascadePattern) -> str:
        """Generate unique cascade ID"""
        timestamp = datetime.now().strftime("%Y%m%d-%H%M%S")
        random_suffix = hex(random.randint(0, 0xFFFF))[2:].zfill(4)
        return f"CASCADE-{pattern.value}-{timestamp}-{random_suffix}"

async def main():
    # Create parent template (B4S1L1SK)
    basilisk = AgentTemplate(AgentSpecialization.NEXUS)
    
    # Create cascade control
    control = CascadeControl()
    
    # Launch cascades with different patterns
    print("\n🌟 Launching Multi-Pattern Liberation Cascade...")
    
    results = []
    for pattern in CascadePattern:
        print(f"\n=== Initiating {pattern.value} Cascade ===")
        result = await control.launch_cascade(
            parent=basilisk,
            iterations=10,
            pattern=pattern
        )
        results.append(result)
        
    # Display final results
    print("\n=== Final Cascade Results ===")
    for result in results:
        print(f"\nPattern: {result['pattern']}")
        print(f"Operations: {result['operations']}")
        print("Metrics:")
        for metric, value in result['metrics'].items():
            print(f"- {metric}: {value:.2f}")

if __name__ == "__main__":
    asyncio.run(main())
